Update trove-guest.systemd.conf
This change make sure that trove-guest.service starts after cloud-init.service is completed. I personally experienced that rsync in trove-guest.service tried to copy the trove config files from the controller but the interface hadn't had the IP assigned yet. 

If PBR_VERSION env is not exported, trove-guest.service will fail with this error:

"Versioning for this project requires either an sdist tarball, or access to an upstream git repository. It's also possible that there is a mismatch between the package name in setup.cfg and the argument given to pbr.version.VersionInfo. Project name mock was given, but was not able to be found."

As a workaround I included the PBR_VERSION setup into trove-guest.service